Add listing of configuration keys
A new view to list the stored configuration keys with a link to load
them in the manager.

This is intended to be used in environments where secrecy is not a
desired feature, as discussed in issue #200. So it is disabled by
default and can be turned on by setting the APPRISE_ALLOW_CONFIG_LIST
environment variable to true.

It will only work when the STATEFUL_MODE is set to simple.